:truck: Flyttet auth til getserversideprops
